# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: BioMinute Publishing Dashboard API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /youtube/publish/{id}:
    post:
      operationId: publishToYouTube
      tags: [youtube]
      summary: Schedule or publish episode to YouTube
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PublishRequest"
      responses:
        "200":
          description: Publish result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PublishResult"
        "400":
          description: Bad request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    PublishRequest:
      type: object
      properties:
        scheduleAt:
          type: ["string", "null"]
          description: ISO datetime to schedule. Null = publish immediately.
        privacyStatus:
          type: string
          enum: [public, unlisted, private]
          default: public
      required: [privacyStatus]
    PublishResult:
      type: object
      properties:
        success:
          type: boolean
        youtubeVideoId:
          type: ["string", "null"]
        youtubeUrl:
          type: ["string", "null"]
        scheduledAt:
          type: ["string", "null"]
        message:
          type: string
      required: [success, message]
...